requirements.txt
# Orchestration
dagster==1.7.0
dagster-webserver==1.7.0
dagster-postgres==0.23.0
dagster-dbt==0.23.0

# Transformation
dbt-core==1.7.0
dbt-postgres==1.7.0

# Data Processing
pandas==2.2.0
numpy==1.26.0
pyarrow==15.0.0
openpyxl==3.1.2

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# API & Data Sources
requests==2.31.0

# Geospatial (optional for mapping)
geopandas==0.14.0
folium==0.15.0

# Visualization & Dashboard
streamlit==1.31.0
plotly==5.18.0
scikit-learn==1.4.0
networkx==3.2.1

# Development
pytest==8.0.0
black==24.1.0
ruff==0.2.0
jupyter==1.0.0
python-dotenv==1.0.0
